fix(core): 对齐会话交接模块结构契约
